+++ service/controller/resource/prometheus/resource.go
@@ -177,10 +177,10 @@ func toPrometheus(v interface{}, config Config) (metav1.Object, error) {
 					corev1.ResourceMemory: *key.PrometheusDefaultMemory(),
 				},
 				Limits: corev1.ResourceList{
-					// cpu: 100m
-					corev1.ResourceCPU: *key.PrometheusDefaultCPU(),
+					// cpu: 150m
+					corev1.ResourceCPU: *key.PrometheusDefaultCPULimit(),
 					// memory: 1.2Gi
-					corev1.ResourceMemory: *key.PrometheusMemoryLimit(),
+					corev1.ResourceMemory: *key.PrometheusDefaultMemoryLimit(),
 				},
 			},
 			Retention:      config.RetentionDuration,